Move units to its own module, imports unchanged
+ Add `Second` as the `Default` for `Unit`.
+ Add `Hertz` as the `Default` for `Freq`.
